DB-Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
